requirements.txt
Flask==2.3.3
textblob==0.17.1
Werkzeug==2.3.7
nltk==3.8.1
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
transformers==4.21.3
torch==2.0.1
flask-cors==4.0.0
python-dotenv==1.0.0